+++ src/main/java/com/hermanvfx/springbackreviewplatform/service/impl/UserServiceImpl.java
@@ -48,7 +48,7 @@ public class UserServiceImpl implements UserService {
 
         return new UserListDto()
                 .content(page.getContent())
-                .totalPages(BigDecimal.valueOf(Math.ceil((double) list.size() / pageable.getPageSize())))
+                .totalPages(BigDecimal.valueOf((int) Math.ceil((double) list.size() / pageable.getPageSize())))
                 .totalElements(BigDecimal.valueOf(list.size()))
                 .currentPage(BigDecimal.valueOf(pageable.getPageNumber()));
     }
